Create OrderTable component for displaying items in cart
