+++ src/components/Checkout/Checkout.js
@@ -2,6 +2,7 @@ import { getProducts } from "../../api/api";
 import React, { useState, useEffect } from "react";
 import LoadingIcon from "../LoadingIcon/LoadingIcon";
 import ProductList from "../ProductList/ProductList";
+import OrderTable from "../OrderTable/OrderTable";
 
 const Checkout = () => {
   const [cart, setCart] = useState({});
@@ -63,50 +64,12 @@ const Checkout = () => {
           </div>
           <div className="checkout-grid">
             <h1>Order Summary</h1>
-            <table className="order-summary-table">
-              <thead>
-                <tr>
-                  <th>Product</th>
-                  <th>Price</th>
-                  <th>Quantity</th>
-                  <th>Total</th>
-                </tr>
-              </thead>
-              <tbody>
-                {Object.keys(cart).map((productId) => {
-                  const product = products.find(
-                    (p) => p.id === parseInt(productId)
-                  );
-                  const quantity = cart[productId];
-                  return (
-                    <tr key={productId}>
-                      <td>{product.name}</td>
-                      <td>${product.price.toFixed(2)}</td>
-                      <td>
-                        <button
-                          onClick={() =>
-                            removeProductFromCart(productId)
-                          }
-                        >
-                          -
-                        </button>
-                        {quantity}
-                        <button
-                          onClick={() =>
-                            addProductToCart(productId)
-                          }
-                        >
-                          +
-                        </button>
-                      </td>
-                      <td>
-                        ${(product.price * quantity).toFixed(2)}
-                      </td>
-                    </tr>
-                  );
-                })}
-              </tbody>
-            </table>
+            <OrderTable
+              cart={cart}
+              products={products}
+              addProductToCart={addProductToCart}
+              removeProductFromCart={removeProductFromCart}
+            />
             <p>Subtotal: ${cartSubtotal.toFixed(2)}</p>
             <p>Discount: ${discount.toFixed(2)}</p>
             <p>Total: ${total.toFixed(2)}</p>
